adjusted to use on CPU

+++ ml_network/models/losses.py
@@ -66,7 +66,7 @@ def NLL_Focal_Loss(
     # Calculate the NLL loss
     if targets.dim() == 2:
         targets = targets.squeeze(1)
-    nll_loss = F.nll_loss(inputs.log(), targets, reduction="none")
+    nll_loss = F.nll_loss(inputs.log(), targets.long(), reduction="none")
 
     # Calculate the Focal Loss
     p_t = inputs[:, 1] * targets + inputs[:, 0] * (1 - targets)
